control-service: fix inifinite redeployment
Why:
The data job synchronizer was stuck in an infinite redeployment loop for all data jobs, occurring every minute. This issue was caused by the random generation of job resources in the cron job definition, where SHA-512 hashing is performed. When we had two instances of the control service running, they generated different SHA codes due to differences in resource order.

What:
We made the following changes to address this issue:

Replaced the resources map with a LinkedHashMap implementation to ensure consistent order on every run.
Reduced the level of logging, as it was overly verbose for one-minute executions.

Testing done:
I conducted local manual testing to ensure the changes function as intended.

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com