vdk-heartbeat: Introduce additional sleep when checking deployments
With the introduction of the deployment synchronizer, the behaviour of the Control Service,
when a data job is deployed, has changed slightly. Now, there is a slight delay between a job
status changing from Enabled to Disabled, and the status being reflected in the output of the
`vdk deploy --show` command. This causes the vdk heartbeat to fail, as it expects the changes
to be applied immediately.

This change adds a new check when the `check_deployments()` method is called, whose purpose is
to verify that the job status has changed from enabled to disabled, and sleep for a bit if this
is not the case.

Testing Done: Executed the heartbeat locally with the command
```bash
vdk-heartbeat -f test_basic_config.ini
```

Signed-off-by: Andon Andonov <andonova@vmware.com>