control-service: Add GraphQL read from DB (#2837)
Currently, when using the GraphQL API, the data job deployment data is
read from the kubernetes
cronjobs. This works fine in general, but it also means that the
kubernetes resources become the
single point of truth, which is not ideal.
    
As part of VEP-2272, the deployment information is moved to the DB,
which improves speed and
reliability. This means that we need a way to read data from the DB,
instead of the kubernetes.
   
This change adds support for reading deployment data from the database
when using the GrapgQL
API. The option to read from the k8s remains.
    
Testing Done: Added test.

---------

Signed-off-by: Andon Andonov <andonova@vmware.com>
Co-authored-by: github-actions <>