vdk-data-sources: add state
Add support for passing and keeping state in data sources.

The way it works a data source in the begining would be called with
method `data_source.connect(previous_state)` to initialize.
Then it if Payload.state field is not None , that state will be
persisted for that stream.

Data Source can keep per-stream state and "others" state for non-stream
specfic stateful information if needed.

auto