vdk-core: create config option for logging execution result
Why?

User reasearch indicates that the execution result is too verbose for local runs. Users don't expect a lot of output for successful jobs and expect error output for failing jobs.

What?

Introduce the LOG_EXECUTION_RESULT config option that enables/disables displaying the end result.
Enabling and disabling the end result will be implemented in a separate PR.
https://github.com/vmware/versatile-data-kit/pull/2831

How was this tested?

CI

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>