vdk-ipython: add %%vdkingest magic function
This commit introduces a new IPython magic function, `vdkingest`, which
enables data ingestion through TOML configuration within a Jupyter
notebook. This enhancement aims to simplify and streamline the data
ingestion process by allowing users to declare their data flows in a
structured format (TOML), thereby making it more accessible and
maintainable.

This is an optional feature of the plugin. If `vdk-data-sources` is not
installed, an ImportError is raised with an informative message.