vdk-jupyter: fix bug for failed requests and improve error handling (#2916)
What: Bug fix. Since the Task Runner change got in, if you tried to
deploy a job without keytab, you would not have been informed about the
error in any way and the Task Runner state wouldn't have been cleaned,
therefore, not being able to handle any other request. Only a restart of
the session would have solved this which is not good.
To address this I:
* Modified the Task Runner start_task() method to proceed after 'failed'
last request.
* Improved the error handling - show the error of failing request which
was not shown and made the backend handlers able to detect any errors.

Why: To make it work in a better way and not to have to restart the
session on each bad request.

Testing Done: Introduced relevant tests. The rest of the tests are
passing.

Signed-off-by: Yoan Salambashev <ysalambashev@vmware.com>