vdk-core: add datetime to decimal json encoder
Why?

As part of the work for vdk-oracle, we need to be able to pass datetime objects
for ingestion. This is not possible with the current verification.

What?

Add datetime to the decimal json encoder
Add bytes to the decimal json encoder
Make the encoder return timestamp for the datetime object
Make the encoder serialize bytes as a list of ints
Rename DecimalJsonEncoder to IngesterJsonEncoder

How was this tested

Ran locally with vdk-oracle
Unit and functional tests
CI tests

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>