vdk-oracle: support type inference
Why?

Inferring the type in case a string is
passed in the payload instead of the corresponding
data type improves the user experience

Type conversion can be relegated to vdk instead of
doing pre-processing of large payloads inside data jobs

What?

Add column data types to column cache
In case a string is passed, infer the python data type based on the Oracle column data type
In case math.nan is passed, write NULL to the Oracle table

How was this tested?

Ran functional tests locally
CI/CD

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>