control-plane: remove needless step in docker build. (#2947)
# Why

Between each step in a docker build it needs to save a snapshot of the
system.
Some operations actually require unpacking layers created before it. 

This python -V is a really expensive operation. 

Because it needs to unpack all the layers that came before it to run
that command.


from the logs we can see 
```
INFO[0001] Unpacking rootfs as cmd RUN python -V requires it.
```

# What
I removed the operation because it actually has no impact. 

# How was this tested? 

I will deploy in CICD env and test when I get approvals. 
I want it removed incase there is some context I am missing.

Co-authored-by: paulm2 <paulm2@vmware.com>