control-service: Add support for vdk version (#2943)
Currently, the vdk image used when a data job is deployed, is decided
based on the python version. This is done in order to avoid possible
incompatibility issues between the vdk image and the job base image.
However, it also blocks the possibility to use testing vdk images in
CI/CD acceptance tests.

This change adds functionality that allows a client to pass a specific
vdk image to be used for a data job deployment. This would allow for
CI/CD tests to test specific vdk images before they are released.

Testing Done: Unit and Integration tests

---------

Signed-off-by: Andon Andonov <andonova@vmware.com>
Co-authored-by: github-actions <>