vdk-impala: handle decorate operation errors (#2975)
## Why?

Due to logic in vdk-core, the managed cursor tries to execute all
implementations of the db_connection_decoreate_operation hook. This
happens even if it's not the correct cursor, e.g. an oracle cursor might
try to execute impala queries. This case is handled in other plugins by
wrapping the decorator in a try/catch block. However, the job should
fail if the default db type is impala and the actual error is an impala
error and not coming from some other cursor.

## What?

Check the db_default_type config if an exception occurs in the
decorator. If the db_default type is not impala, output an error log. If
the db_default type is impala, re-throw the error

## How was this tested?

Tested locally
CI

## What kind of change is this?

Bugfix

## Follow-up

https://github.com/vmware/versatile-data-kit/issues/2973

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>