vdk-control-cli: Modify how data is written to job config.ini
Currently, the `_set_value()` logic in **JobConfig** uses `fileinput.input()` to traverse the contents of
the job config file, and update the keys passed. This works as expected on python versions up to 3.11. It,
however breaks on python3.11 on some platforms. For example, it works fine on Ubuntu, but the config file
gets truncated on photonOS (the operating system used for the secure base images), which causes job creation
failures.

This change modifies the `_set_value()` logic to directly read and write the config file's contents, and drops
the use of *fileinput*.

Testing Done: Added test, and did some local testing using docker containers with different host OSs.

Signed-off-by: Andon Andonov <andonova@vmware.com>