Update to READMEs: Add a monthly download badge
Why?
Allow users to get a brief overview of popularity of the project or plugin

What?
A monthly download badge on each Readme.

How has this been tested?
Each badge has been checked on GitHub and the links to PyPI have been checked to go to the right package.

What type of change are you making?
Cosmetic/ Documentation