vdk-oracle: fixes
1. Fix the tests to be timezone independant. Currently they were failing
because fromtimestamp is timezone dependant so it will return different
reseult depending on the timezone

2. Simplify passing secrets as configuration by removing redudant
configuration options and note it's just a prototype until such change
is in core. It is unnecsary to pass 3 configuration options to start
passing secret. Just read everytime secrets if there's one use if not
no. Kept option "ORACLE_USE_SECRETS" as a kill switch just in case.