vdk-oracle: add new config options (host,port,sid,thick_mode_lib_dir)
We are adding 3 options host,port,sid that were being originally added
as part of PR https://github.com/vmware/versatile-data-kit/pull/3019

The issue in the PR is that SID was incorrecctly passed as
"service_name" to the underlying oracle db connect method. But it is
actuall "SID" (system identifier) .

Setting it correctly there works fine.

I also had to add thick_mode_lib_dir since it was impossible to get
thick mode working in Mac without passing it on.

Testing Done:
	- added   test_oracle_connect_execute_without_connection_string
	- tested in production environment against real (non local) with
thick mode database with a data job which executed query and send data
for ingestion successfully.

Currently the automated tests do not cover the thick mode scenario due
to extra setup required and this will be added later.