vdk-core: remove structlog logging override (#3066)
## Why?

Structlog should not override the initial logging config. It applies to
logging before vdk is initialized and before any plugins are loaded. If
an error happens in vdk_main, the logging config will apply. It should
remain in core.

It's also not a problem for structlog, because it ejects all logging
config just in case.

## What?

Remove the structlog override in core

## How was this tested?

Locally
CI

## What kind of change is this?

Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>
Co-authored-by: Dilyan Marinov <mdilyan@vmware.com>