vdk-dag: fix unnecessary authorization failure
If you have locally configured authorizaton settings (e.g client id,
refresh token and so on) which might be expired or wrong, VDK dag still
tries to authenticate and generate access token even if no
authentication may be required by the Control Service API. This casues
the DAG jobs to fail with authentication error which is unnecssary if
the API didn't require auth.

Instead we now just post a warning about the failure to get acess token.
We fail later if and only if the API requires authentication . If not
everything suceeeds.

Testing Done: local tests pass (they failed before for above reason)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>