RAG Question-answering Web Service
This PR introduces a script which  allows for the quick deployment of
a local service which can answer questions based on the context
provided from a configured vector database.

I've left the table names hardcoded, looking for suggestions as to
what the best approach for making them configurable is - putting them
in the config.ini file did not seem appropriate.

Testing done: local

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>