vdk-core: enable/disable structlog based on config
Why?

vdk-structlog has a default config flag set to True when installed.
It relegates all logging configuration to the plugin and is cheked in core,
i.e. if structlog is installed, vdk will use structlog. If it's not installed,
it will use the configuration available in core. The flag is currently checked
using an env variable, which makes it not work.

What?

Check the flag using the config object

How was this tested?

Manually

What kind of change is this?

Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>