vdk-frontend: Show Python Version of the Data Jobs Deployment(#SC-2456)
As a DE who needs to change the source code of a job, knowing actual Python version
with which a data job has been last deployed makes it easier.
Also, knowing historically with what Python version used for the deployment of this data job for execution.

-We now showcase the python version used for a data-job deployment in the data-job definition
-The history of data-job executions also now displays the python-version used for deployment

Local tests passed

Feature/ Non-breaking