vdk-test-utils: Measure payload size with len, not getsizeof (#3157)
Previously, the size of payloads was measured using sys.getsizeof when
testing. This is suboptimal, since different versions of Python might
give different sizes (and in particular, when using Python3.12, the size
of payloads is 8 bytes less than in previous versions). To address this,
we now use len to measure the payload size, which should be consistent
across all Python versions.

The relevant tests in vdk-core have been adjusted to address this
change.

Testing done: cicd

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>