Control-service: load iam role eks service account (#3201)
# What 

When running on AWS and we need to access the private amazon container
registry we make use of IAM roles to give us permission to access this
resource.

When using an IAM role you can only get an access token that lasts a
couple of hours and then you need to refresh it.
The class `AWSCredentialsService` is responsible for doing this refresh.


But this leads to a chicken end egg situation. 
the question how can you make a request for an IAM access token? 
Surely that requires authenticating yourself too. 

Yes, it does. 
So within our existing code you need to pass in the access token and
secret key of a user who has permission to take on this IAM role.

However when a job is running on an EKS cluster it doesn't need to do
that.
We mark all pods on this eks cluster as being able to use this IAM role.

But to support that we need to create the token refresher using
`DefaultAWSCredentialsProviderChain` instead of using
`AWSStaticCredentialsProvider`.

In this PR I check if the passed in access token and secret is empty if
it is then I use the `DefaultAWSCredentialsProviderChain` which will
work on AWS.


# How was this tested?
testing on an EKS cluster everything works perfectly.

---------

Co-authored-by: github-actions <>
Co-authored-by: paulm2 <paulm2@vmware.com>