vdk-core: support overriding configs with secrets (#3125)
## Why?

VDK doesn't provide a way to set sensitive configuration like passwords,
such as trino_password. The only way to currently do this is by adding
config keys and fetching the values from secrets.

## What?

Add a plugin that reconfigures the Configuration object in CoreContext
based on secrets. Do this in the initialize_job hook. In this setup,
secrets override options set by regular configs. For example if you set
trino_password to "password" in config.ini, but also have a secret
called trino_passowrd="another password", the value of trino_password
will be "another_passowrd".

**Note:** The Configuration class and the CoreContext class are
annotated with `@dataclass(frozen=True)`. This enforces encapsulation,
so in order to mutate the Configuration object after it's created, we
have to add more public methods to the Configuration class.

https://docs.python.org/3/library/dataclasses.html#frozen-instances

## How was this tested?

Functional test
CI/CD

## What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>