control-service: remove support for old k8s (#3202)
# Why 
External customers are looking to run on newer versions of k8s than we
support.
We are using a very old batch/V1Beta1 api. 
We need to remove all references to this.

# What 
I deleted all code which was referencing it. This looks like a massive
PR but it should actually be fairly simple to review because it is just
deleting code.
 
# How was this tested??
Existing unit and integration tests. 
I will also be deploying it to newer versions of k8s this week.



Signed-off-by: murphp15 <murphp15@tcd.ie>

---------

Co-authored-by: github-actions <>
Co-authored-by: paulm2 <paulm2@vmware.com>