control-service: simplify deployment of VDK helm chart (#3214)
# Why 
When VDK helm chart is deployed without ingress setup the UI is actually
unusable.
It doesn't pull information from the control-plane and so is just an
empty UI.

This is because the frontend relies on a specific ingress setup to route
traffic to the control-plane.

Historically we have expected ppl to create this ingress route them.
However this is stupid.
Because its not obvious how this needs to be setup. 


In this PR I include a ingress controller in the chart and include
default config for ingress which would create an appropriate ingress
controller for the fe and be to be able to work together.

**Important**
I still keep it disabled by default because we need to know the host
name before we can create the rule.
But no once they have the host name the work is trivial to complete. 

# How has this been tested?
I deployed many times

---------

Co-authored-by: paulm2 <paulm2@vmware.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>