vdk-core: add more restrictive hooks to hookspec
Why?

The hooks that expose cursors end up in pluggy's
LIFO buffer and execute every time. This is ok for
things like lineague or metrics, but it breaks
database plugins if they try to execute queries
with the wrong cursor.

What?

Create more restrictive hooks to eventually replace the hooks
that expose the DecorationCursor and Recovery cursor

How was this tested?

Functional tests

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>