vdk-oracle: match table columns to payload row keys
Why?

Fetching the table columns from the database returns them in
all-caps, unless case-sensitive is set in the database. This
can happen either by enabling it for the whole db or by putting
identifiers in quotes. We need to take this into account when
ingesting

What?

For every column, check if the column is in the payload.
If it's not in the payload, check if the upper-case or
lower-case column identifier is in the payload. This is
the best we can do if the database is case-insensitive.

How was this tested?

Functional tests

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>