control-service: add amazon ses smtp mail notifications (#3269)
# Why
We want to be able to send email notifications using Amazon SES SMTP. 

# What
Until now, we only had default SMTP configuration for notifications. We
move this logic to a new DefaultEmailNotification class which implements
the EmailNotification interface. We also introduce an Amazon-oriented
AmazonSesEmailNotification class which also implements the
EmailNotification interface.

# How was this tested
It was just SMTP code until now. 
It was tested in an internal k8s cluster as we don't have a present
testing SES.


Signed-off-by: Yoan Salambashev <yoan.salambashev@broadcom.com>

---------

Signed-off-by: Yoan Salambashev <yoan.salambashev@broadcom.com>
Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>
Signed-off-by: Duygu Hasan [hduygu@vmware.com](mailto:hduygu@vmware.com)
Co-authored-by: Yoan Salambashev <yoan.g.salambashev@gmail.com>
Co-authored-by: github-actions <>
Co-authored-by: Duygu Hasan <87015481+duyguHsnHsn@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dilyan Marinov <91800778+DeltaMichael@users.noreply.github.com>
Co-authored-by: Dilyan Marinov <mdilyan@vmware.com>