control-service: rewrite correctly for spies (#3286)
# Why
The previous code is not written assuming assuming the object in a mock
when infact it is a spy.
The difference between a mock and a spy is that with a spy the real
functions are called.

That is why we can remove the line with `.callRealMethod`.
I have also rewritten how we return mocked functions response objects. 
They way they were written was resulting in the actual function being
called when we are first mocking it.

However now it never calls it. 


I didn't rename it from mock to spy. because this is everywhere in the
code and can be done in a different PR if ppl think it is important.




Tested on CICD