control-service: errors should be correctly propagated. (#3263)
# Why
At the moment if the kubernetes client throws and error then we assume
that the job is not there.
This is obviously not correct. We should only consider the job not there
if it is a 404.

# What
Now we check to see the status to decide what to do. 

# How was this tested
New unit test which mock k8s responses.


Signed-off-by: murphp15 <murphp15@tcd.ie>

---------

Co-authored-by: github-actions <>