vdk-server: fix the local installation (#3300)
What: Fix the unstable vdk-server. The CockroachDB was causing problems
in the chart which led to the decision to switch to Postgres. This
change updates the postgresql chart and uses it. Also, there was a
hanging problem with the type of the ingress nginx controller service so
I switched it to ClusterIP.

Why: To have a working local installation of VDK Control Service.

Testing Done: vdk server -i

Signed-off by: Yoan Salambashev <yoan.salambashev@broadcom.com>

---------

Co-authored-by: Yoan Salambashev <yoan.g.salambashev@gmail.com>