vdk-trino: Create scd_upsert template
What:
Adding new way to process data. It is meant to update target table rows based on id column that are coming from the source view and also insert the new rows that are not present in the target table already.

Why:
It will be used mainly for optimizing the scd1 template in order to avoid calculating the whole table logic on every run.